+++ packages/thumbprint-react/components/Image/index.tsx
@@ -71,6 +71,7 @@ type ObjectFitPropsType = {
         objectFit?: 'cover' | 'contain';
         objectPosition?: 'top' | 'center' | 'bottom' | 'left' | 'right';
         fontFamily?: React.CSSProperties['fontFamily'];
+        height?: '100%';
     };
 };
 
@@ -139,7 +140,10 @@ const Image = forwardRef<HTMLElement, ImagePropTypes>((props: ImagePropTypes, ou
 
     const objectFitProps: ObjectFitPropsType = {};
 
-    const shouldObjectFit = !!height;
+    // Checking for the use of the `height` prop is not enough since users can also change the
+    // image height using `className`, or `style`.
+    const shouldObjectFit = !!height || !!props.objectFit;
+
     const shouldPolyfillObjectFit =
         canUseDOM &&
         document.documentElement &&
@@ -170,6 +174,18 @@ const Image = forwardRef<HTMLElement, ImagePropTypes>((props: ImagePropTypes, ou
             objectFit,
             objectPosition,
         };
+
+        if (!height) {
+            // Add `height: 100%` as an inline style if the user wants to `objectFit` but hasn't
+            // passed in the `height` prop. Almost always, this means that the user is setting the
+            // height with CSS or an inline style. Since inline styles and `className` get added to
+            // `picture`, not `img`, the `img` element would become taller than the picture,
+            // preventing the `objectFit` from working. Adding `height: 100%` to the `img` in these
+            // cases allows `objectFit` to work as well as it would if the `height` was provided as
+            // a prop rather than through `style` or `className`.
+            objectFitProps.style.height = '100%';
+        }
+
         if (shouldPolyfillObjectFit) {
             // Weird, but this is how the polyfill knows what to do with the image in IE.
             objectFitProps.style.fontFamily = `"object-fit: ${objectFit}; object-position: ${objectPosition}"`;